requirements.txt
opencv-python
mediapipe
numpy

# For detect people
supervision
ultralytics

# For transformers based, dynamic no training detection
transformers 
torchvision 
matplotlib

# For lazy frontend
gradio